修正了圆角矩形的混合问题，现在圆角边框不在有深色的情况了

+++ CAI_tempName/src/Controls/Shap/Path.cpp
@@ -1,7 +1,7 @@
  #include "caipch.h"
 #include "Path.h"
 
-void Path::AddPoint(const double& x, const double& y)
+void Path::AddPoint(const float& x, const float& y)
 {
 	if (y > height.get())
 		height.set(y);
