+++ CAI_tempName/src/Datas/Coordinate.h
@@ -42,7 +42,7 @@ private:
 class Size : public Object {
 public:
 	Size()noexcept :resolutionWidth(0.0),resolutionHeight(0.0),x(0.0), y(0.0), z(0.0), height(0.0), width(0.0) {}
-	Size(const double& x, const double& y, const double& z, const double& width, const double& height) noexcept :resolutionWidth(0.0), resolutionHeight(0.0),x(x), y(y), z(z), height(height), width(width) {}
+	Size(const float& x, const float& y, const float& z, const float& width, const float& height) noexcept :resolutionWidth(0.0), resolutionHeight(0.0),x(x), y(y), z(z), height(height), width(width) {}
 	~Size() = default;
 	Size(const Size& other) noexcept {
 		*this = other;
@@ -50,34 +50,34 @@ public:
 public:
 	
 
-	inline double& X() { return x; }
-	inline double& Y() { return y; }
-	inline double& Z() { return z; }
-	inline double& Width() { return width; }
-	inline double& Height() { return height; }
-	inline double& ResolutionWidth() { return resolutionWidth; }
-	inline double& ResolutionHeight() { return resolutionHeight; }
+	inline float& X() { return x; }
+	inline float& Y() { return y; }
+	inline float& Z() { return z; }
+	inline float& Width() { return width; }
+	inline float& Height() { return height; }
+	inline float& ResolutionWidth() { return resolutionWidth; }
+	inline float& ResolutionHeight() { return resolutionHeight; }
 	inline Math::mat4& ModelMatrix() { return mode; }
 	inline Math::mat4& ProjectionMatrix() { return projection; }
 
-	inline const double& X() const { return x; }
-	inline const double& Y() const { return y; }
-	inline const double& Z() const { return z; }
-	inline const double& Width()  const { return width; }
-	inline const double& Height() const { return height; }
-	inline const double& ResolutionWidth() const { return resolutionWidth; }
-	inline const double& ResolutionHeight() const { return resolutionHeight; }
+	inline const float& X() const { return x; }
+	inline const float& Y() const { return y; }
+	inline const float& Z() const { return z; }
+	inline const float& Width()  const { return width; }
+	inline const float& Height() const { return height; }
+	inline const float& ResolutionWidth() const { return resolutionWidth; }
+	inline const float& ResolutionHeight() const { return resolutionHeight; }
 	inline const  Math::mat4& ModelMatrix() const { return mode; }
 	inline const Math::mat4& ProjectionMatrix() const { return projection; }
 
-	inline void SetX(const double& value) { x = value; }
-	inline void SetY(const double& value) { y = value; }
-	inline void SetZ(const double& value) { z = value; }
-	inline void SetXYZ(const double& x, const double& y, const double& z) { this->x = x; this->y = y; this->z = z; }
-	inline void SetWidth(const double& value) { width = value; }
-	inline void SetHeight(const double& value) { height = value; }
-	inline void SetWH(const double& width, const double& height) { this->width = width; this->height = height; }
-	inline void SetResolution(const double& x, const double& y) { resolutionWidth = x; resolutionHeight = y; }
+	inline void SetX(const float& value) { x = value; }
+	inline void SetY(const float& value) { y = value; }
+	inline void SetZ(const float& value) { z = value; }
+	inline void SetXYZ(const float& x, const float& y, const float& z) { this->x = x; this->y = y; this->z = z; }
+	inline void SetWidth(const float& value) { width = value; }
+	inline void SetHeight(const float& value) { height = value; }
+	inline void SetWH(const float& width, const float& height) { this->width = width; this->height = height; }
+	inline void SetResolution(const float& x, const float& y) { resolutionWidth = x; resolutionHeight = y; }
 	bool PointIn(const float& x, const float& y) const {
 		return (x >= this->x && x <= this->x + width) && (y >= this->y && y <= this->y + height);
 	}
@@ -95,17 +95,17 @@ public:
 		return *this;
 	}
 private:
-	double resolutionWidth;
-	double resolutionHeight;
+	float resolutionWidth;
+	float resolutionHeight;
 	//局部坐标X，空间左上角
-	double x;
+	float x;
 	//局部坐标Y，空间左上角
-	double y;
-	double z;
+	float y;
+	float z;
 	//可视化的空间宽度
-	double width;
+	float width;
 	//可视化的空间高度
-	double height;
+	float height;
 	//模型矩阵--用于将自身的坐标转换成全局坐标
 	Math::mat4 mode;
 	//投影矩阵
